+++ backend/private-graph/graph/auth_route.go
@@ -3,18 +3,14 @@ package graph
 import (
 	"encoding/json"
 	"errors"
-	"github.com/highlight-run/highlight/backend/env"
-	"net/http"
-	"time"
-
 	"firebase.google.com/go/auth"
-
-	"github.com/golang-jwt/jwt/v4"
+	"github.com/highlight-run/highlight/backend/env"
+	"github.com/highlight-run/highlight/backend/model"
 	e "github.com/pkg/errors"
 	"github.com/sendgrid/sendgrid-go/helpers/mail"
 	log "github.com/sirupsen/logrus"
-
-	"github.com/highlight-run/highlight/backend/model"
+	"net/http"
+	"time"
 )
 
 type LoginCredentials struct {
@@ -24,62 +20,35 @@ type LoginCredentials struct {
 
 var AdminPassword = env.Config.AuthAdminPassword
 var AdminPasswordTokenDuration = time.Hour * 24
+var LoginError = "invalid email/password provided"
+var LoginFlowError = "login flow not supported"
 
 func (r *Resolver) Login(w http.ResponseWriter, req *http.Request) {
-	var credentials LoginCredentials
-
 	ctx := req.Context()
 
-	if AdminPassword == "" {
-		http.Error(w, "", http.StatusInternalServerError)
-		log.WithContext(ctx).Error(errors.New("Password auth mode not properly configured."))
-		return
-	}
-
+	var credentials LoginCredentials
 	err := json.NewDecoder(req.Body).Decode(&credentials)
-
-	if err != nil || credentials.Password == "" || credentials.Email == "" {
-		log.WithContext(ctx).Error(err)
-		http.Error(w, "no email or password provided", http.StatusBadRequest)
-		return
+	if err != nil {
+		log.WithContext(ctx).WithError(err).Error("failed to unmarshal login details")
+		http.Error(w, LoginError, http.StatusInternalServerError)
 	}
 
 	_, err = mail.ParseEmail(credentials.Email)
-
 	if err != nil {
-		log.WithContext(ctx).Error(err)
-		http.Error(w, "Invalid email provided", http.StatusBadRequest)
-		return
+		log.WithContext(ctx).WithError(err).Error("failed to parse login email")
+		http.Error(w, LoginError, http.StatusInternalServerError)
 	}
 
-	if AdminPassword != credentials.Password {
-		http.Error(w, "invalid password", http.StatusBadRequest)
-		return
+	response, err := AuthClient.PerformLogin(ctx, credentials)
+	if err != nil {
+		log.WithContext(ctx).WithError(err).Error("failed to validate login")
+		http.Error(w, LoginError, http.StatusInternalServerError)
 	}
 
-	user := GetPasswordAuthUser(credentials.Email)
-
-	atClaims := jwt.MapClaims{}
-	atClaims["authorized"] = true
-	atClaims["exp"] = time.Now().Add(AdminPasswordTokenDuration).Unix()
-	atClaims["email"] = user.Email
-	atClaims["uid"] = user.Email
-	at := jwt.NewWithClaims(jwt.SigningMethodHS256, atClaims)
-
-	token, err := at.SignedString([]byte(JwtAccessSecret))
+	jsonResponse, err := json.Marshal(response)
 	if err != nil {
 		log.WithContext(ctx).Error(err)
-		http.Error(w, "", http.StatusInternalServerError)
-	}
-
-	response := make(map[string]interface{})
-	response["token"] = token
-	response["user"] = user
-
-	jsonResponse, err := json.Marshal((response))
-	if err != nil {
-		log.WithContext(ctx).Error(err)
-		http.Error(w, "", http.StatusInternalServerError)
+		http.Error(w, LoginError, http.StatusInternalServerError)
 	}
 
 	w.Header().Set("Content-Type", "application/json")
@@ -98,7 +67,7 @@ func (r *Resolver) ValidateAuthToken(w http.ResponseWriter, req *http.Request) {
 	uid := ctx.Value(model.ContextKeys.UID)
 
 	if email == nil || uid == nil {
-		log.WithContext(ctx).Error(errors.New("Forbidden"))
+		log.WithContext(ctx).Error(errors.New("forbidden"))
 		http.Error(w, "", http.StatusUnauthorized)
 		return
 	}
@@ -108,7 +77,7 @@ func (r *Resolver) ValidateAuthToken(w http.ResponseWriter, req *http.Request) {
 	response := make(map[string]interface{})
 	response["user"] = user
 
-	jsonResponse, err := json.Marshal((response))
+	jsonResponse, err := json.Marshal(response)
 	if err != nil {
 		log.WithContext(ctx).Error(err)
 		http.Error(w, "", http.StatusInternalServerError)
