+++ backend/private-graph/graph/middleware.go
@@ -9,24 +9,13 @@ import (
 	"io"
 	"net/http"
 	"strings"
-	"time"
-
-	firebase "firebase.google.com/go"
-	"firebase.google.com/go/auth"
-	"golang.org/x/oauth2/google"
-	"google.golang.org/api/option"
-
-	"github.com/golang-jwt/jwt/v4"
-	"github.com/samber/lo"
-	log "github.com/sirupsen/logrus"
 
 	"github.com/99designs/gqlgen/graphql/handler/transport"
 	"github.com/go-oauth2/oauth2/v4"
-	e "github.com/pkg/errors"
-
 	"github.com/highlight-run/highlight/backend/model"
 	"github.com/highlight-run/highlight/backend/oauth"
 	"github.com/highlight-run/highlight/backend/util"
+	log "github.com/sirupsen/logrus"
 )
 
 type APITokenHandler func(ctx context.Context, apiKey string) (*int, error)
@@ -45,6 +34,8 @@ const (
 	Simple   AuthMode = "Simple"
 	Firebase AuthMode = "Firebase"
 	Password AuthMode = "Password"
+	LDAP     AuthMode = "LDAP"
+	SAML     AuthMode = "SAML"
 )
 
 func GetEnvAuthMode() AuthMode {
@@ -57,140 +48,22 @@ func GetEnvAuthMode() AuthMode {
 	return Firebase
 }
 
-type Client interface {
-	updateContextWithAuthenticatedUser(ctx context.Context, token string) (context.Context, error)
-	GetUser(ctx context.Context, uid string) (*auth.UserRecord, error)
-}
-
-type SimpleAuthClient struct{}
-
-type PasswordAuthClient struct{}
-
-type FirebaseAuthClient struct {
-	AuthClient *auth.Client
-}
-
-func (c *FirebaseAuthClient) GetUser(ctx context.Context, uid string) (*auth.UserRecord, error) {
-	return c.AuthClient.GetUser(ctx, uid)
-}
-
 func SetupAuthClient(ctx context.Context, authMode AuthMode, oauthServer *oauth.Server, wsTokenHandler APITokenHandler) {
 	OAuthServer = oauthServer
 	workspaceTokenHandler = wsTokenHandler
 	if authMode == Firebase {
-		secret := env.Config.AuthFirebaseSecret
-		creds, err := google.CredentialsFromJSON(ctx, []byte(secret),
-			"https://www.googleapis.com/auth/firebase",
-			"https://www.googleapis.com/auth/identitytoolkit",
-			"https://www.googleapis.com/auth/userinfo.email")
-		if err != nil {
-			log.WithContext(ctx).Errorf("error converting credentials from json: %v", err)
-			return
-		}
-		app, err := firebase.NewApp(context.Background(), nil, option.WithCredentials(creds))
-		if err != nil {
-			log.WithContext(ctx).Errorf("error initializing firebase app: %v", err)
-			return
-		}
-		// create a client to communicate with firebase project
-		var client *auth.Client
-		if client, err = app.Auth(context.Background()); err != nil {
-			log.WithContext(ctx).Errorf("error creating firebase client: %v", err)
-			return
-		}
-		AuthClient = &FirebaseAuthClient{AuthClient: client}
+		AuthClient = NewFirebaseClient(ctx)
 	} else if authMode == Simple {
 		AuthClient = &SimpleAuthClient{}
 	} else if authMode == Password {
 		AuthClient = &PasswordAuthClient{}
+	} else if authMode == LDAP {
+		AuthClient = NewLDAPClient(ctx)
+	} else if authMode == SAML {
+		AuthClient = NewSAMLClient(ctx)
 	} else {
 		log.WithContext(ctx).Fatalf("private graph auth client configured with unknown auth mode")
 	}
-}
-
-func (c *SimpleAuthClient) GetUser(_ context.Context, uid string) (*auth.UserRecord, error) {
-	return &auth.UserRecord{
-		UserInfo:      GetPasswordAuthUser(uid),
-		EmailVerified: true,
-	}, nil
-}
-
-func authenticateToken(tokenString string) (jwt.MapClaims, error) {
-	claims := jwt.MapClaims{}
-	_, err := jwt.ParseWithClaims(tokenString, claims, func(token *jwt.Token) (interface{}, error) {
-		return []byte(JwtAccessSecret), nil
-	})
-	if err != nil {
-		return claims, e.Wrap(err, "invalid id token")
-	}
-
-	exp, ok := claims["exp"]
-	if !ok {
-		return claims, e.Wrap(err, "invalid exp claim")
-	}
-
-	expClaim := int64(exp.(float64))
-	if time.Now().After(time.Unix(expClaim, 0)) {
-		return claims, e.Wrap(err, "token expired")
-	}
-
-	return claims, nil
-}
-
-func (c *PasswordAuthClient) GetUser(_ context.Context, uid string) (*auth.UserRecord, error) {
-	return &auth.UserRecord{
-		UserInfo:      GetPasswordAuthUser(uid),
-		EmailVerified: true,
-	}, nil
-}
-
-func (c *PasswordAuthClient) updateContextWithAuthenticatedUser(ctx context.Context, token string) (context.Context, error) {
-	var uid string
-	email := ""
-
-	if token != "" {
-		claims, err := authenticateToken(token)
-		if err != nil {
-			return ctx, err
-		}
-
-		email = claims["email"].(string)
-		uid = claims["uid"].(string)
-	}
-
-	ctx = context.WithValue(ctx, model.ContextKeys.UID, uid)
-	ctx = context.WithValue(ctx, model.ContextKeys.Email, email)
-	return ctx, nil
-}
-
-func (c *SimpleAuthClient) updateContextWithAuthenticatedUser(ctx context.Context, token string) (context.Context, error) {
-	ctx = context.WithValue(ctx, model.ContextKeys.UID, "demo@example.com")
-	ctx = context.WithValue(ctx, model.ContextKeys.Email, "demo@example.com")
-	return ctx, nil
-}
-
-func (c *FirebaseAuthClient) updateContextWithAuthenticatedUser(ctx context.Context, token string) (context.Context, error) {
-	var uid string
-	email := ""
-	if token != "" {
-		t, err := c.AuthClient.VerifyIDToken(context.Background(), token)
-		if err != nil {
-			return ctx, e.Wrap(err, "invalid id token")
-		}
-		uid = t.UID
-		if userRecord, err := c.AuthClient.GetUser(context.Background(), uid); err == nil {
-			email = userRecord.Email
-
-			// This is to prevent attackers from impersonating Highlight staff.
-			_, isAdmin := lo.Find(HighlightAdminEmailDomains, func(domain string) bool { return strings.Contains(email, domain) })
-			if isAdmin && !userRecord.EmailVerified {
-				email = ""
-			}
-		}
-	}
-	ctx = context.WithValue(ctx, model.ContextKeys.UID, uid)
-	ctx = context.WithValue(ctx, model.ContextKeys.Email, email)
-	return ctx, nil
 }
 
 func getSourcemapRequestToken(r *http.Request) string {
